Firmware read checks + integer data

+++ src/firmware/pru_bridge.h
@@ -5,8 +5,8 @@
 
 
 int check_init(void);
-uint8_t read_buffer(int ring_no);
-void write_buffer(int ring_no,uint8_t data);
+uint16_t read_buffer(int ring_no);
+void write_buffer(int ring_no,uint16_t data);
 void pru_bridge_init(void);
 
 #endif
